ex 4.15-4.22

+++ part4/blog list/controllers/blogs.js
@@ -1,26 +1,39 @@
 const blogRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogRouter.get('/', async (req, res) => {
-	const blogs = await Blog.find({})
+	const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
 	res.json(blogs.map((blog) => blog.toJSON()))
 })
 
 blogRouter.post('/', async (req, res) => {
 	const body = req.body
 
-	if (body.title === undefined || body.url === undefined) {
-		return res.status(400).end()
+	const decodedToken = jwt.verify(req.token, process.env.SECRET)
+
+	if (!req.token || !decodedToken.id) {
+		return res.status(401).json({ error: 'token missing or invalid' })
 	}
+	const user = await User.findById(decodedToken.id)
 
 	const blog = new Blog({
 		title: body.title,
 		author: body.author,
 		url: body.url,
 		likes: body.likes === undefined ? 0 : body.likes,
+		user: user._id,
 	})
 
+	if (body.title === undefined || body.url === undefined) {
+		return res.status(400).end()
+	}
+
 	const savedBlog = await blog.save()
+
+	user.blogs = user.blogs.concat(savedBlog._id)
+	await user.save()
 
 	res.status(201).json(savedBlog)
 })
@@ -42,6 +55,24 @@ blogRouter.put('/:id', async (req, res) => {
 })
 
 blogRouter.delete('/:id', async (req, res) => {
+	const blog = await Blog.findById(req.params.id)
+
+	const decodedToken = jwt.verify(req.token, process.env.SECRET)
+
+	if (!req.token || !decodedToken.id) {
+		return res.status(400).json({ error: 'token missing or invalid' })
+	}
+
+	if (blog.user.toString() !== decodedToken.id) {
+		return res.status(403).json({ error: 'unauthorized user' })
+	}
+
+	const user = await User.findById(blog.user)
+
+	user.blogs = user.blogs.filter((blog) => blog.toString() !== req.params.id)
+
+	await user.save()
+
 	await Blog.findByIdAndRemove(req.params.id)
 	res.status(204).end()
 })
